+++ brainbox/tests/test_singlecell.py
@@ -30,7 +30,7 @@ class TestPeths(unittest.TestCase):
         event_times = np.sort(np.random.rand(n_events, ) * record_length)
 
         peth, fr = calculate_peths(spike_times, spike_clusters, cluster_ids=cluster_sel,
-                         align_times=event_times)
+                                   align_times=event_times)
         self.assertTrue(peth.means.shape[0] == len(cluster_sel))
         self.assertTrue(np.all(peth.means.shape == peth.stds.shape))
         self.assertTrue(np.all(fr.shape == (n_events, len(cluster_sel), 28)))