+++ tests/ibllib/test_pipes.py
@@ -3,6 +3,7 @@ import tempfile
 import unittest
 from pathlib import Path
 
+import alf.io
 import ibllib.pipes.extract_session
 from ibllib.pipes import experimental_data, misc
 
@@ -49,7 +50,7 @@ class TestExtractors(unittest.TestCase):
         out = (Path('/mnt/s0/Data/Subjects/ZM_1368/2019-04-19/001'),
                Path('/mnt/s0/Data/Subjects/ZM_1368/2019-04-19/001'),)
         for i, o in zip(inp, out):
-            self.assertEqual(o, ibllib.pipes.extract_session.get_session_path(i))
+            self.assertEqual(o, alf.io.get_session_path(i))
 
 
 class TestPipesMisc(unittest.TestCase):
@@ -244,6 +245,36 @@ class TestPipesMisc(unittest.TestCase):
             '_spikeglx_ephysData_g0_t0.nidq.meta' in nidq_file_names
         )
 
+    def test_create_ephys_flags(self):
+        extract = self.local_session_path_3A.joinpath('extract_ephys.flag')
+        qc = self.local_session_path_3A.joinpath('raw_ephys_qc.flag')
+        spike_sorting0 = self.local_session_path_3A / 'raw_ephys_data' / 'probe00'
+        spike_sorting1 = self.local_session_path_3A / 'raw_ephys_data' / 'probe01'
+        spike_sorting0 = spike_sorting0.joinpath('spike_sorting.flag')
+        spike_sorting1 = spike_sorting1.joinpath('spike_sorting.flag')
+        misc.create_ephys_flags(self.local_session_path_3A)
+        self.assertTrue(extract.exists())
+        self.assertTrue(qc.exists())
+        self.assertTrue(spike_sorting0.exists())
+        self.assertTrue(spike_sorting1.exists())
+        # Test recreate
+        misc.create_ephys_flags(self.local_session_path_3A)
+        self.assertTrue(extract.exists())
+        self.assertTrue(qc.exists())
+        self.assertTrue(spike_sorting0.exists())
+        self.assertTrue(spike_sorting1.exists())
+        # Remove flags after test
+        extract.unlink()
+        qc.unlink()
+        spike_sorting0.unlink()
+        spike_sorting1.unlink()
+        # test removal
+        self.assertFalse(extract.exists())
+        self.assertFalse(qc.exists())
+        self.assertFalse(spike_sorting0.exists())
+        self.assertFalse(spike_sorting1.exists())
+
+# TODO: test copy wirings
     def tearDown(self):
         shutil.rmtree(self.root_test_folder, ignore_errors=True)